requirements.txt
# CaRec
pyepics
# CaPlay
pycaspy
